- added Compatibility folder to Stl.Net; - marked compatibility as internal;
